requirements.txt
langchain
openai
feedparser
yfinance
python-dotenv 